email_system/processor/requirements.txt
kafka-python
python-dotenv  # Opcional para manejar variables de entorno